fix: address review — consolidate port validation, fix uninstall.sh scoping bug
1. Consolidate port validation: extract parsePort into nemoclaw/src/lib/ports.ts
   and import DASHBOARD_PORT from there in runner.ts, eliminating the duplicated
   parseDashboardPort() function.

2. Fix uninstall.sh scoping bug: move _dp assignment above the while loop so
   pgrep's process substitution sees the validated port value. Previously _dp
   was declared inside the loop body but referenced in the process substitution
   which is evaluated before the loop runs, causing pgrep to match ALL
   openshell forward processes regardless of port.
